requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==43.0.3
cssbeautifier==1.15.1
defusedxml==0.7.1
Django==4.2.16
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-resized==1.0.3
django-summernote==0.8.20.0
django-tinymce==4.1.0
EditorConfig==0.12.4
idna==3.10
jsbeautifier==1.15.1
json5==0.9.25
oauthlib==3.2.2
pathspec==0.12.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.67.0
urllib3==2.2.3
webencodings==0.5.1
